pyproject.toml
[tool.poetry]
name = "chipmunk"
version = "0.1.0"
description = "Asset management and portfolio app"
authors = ["benjamin <benjaminhon86@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
pandas = "^1.2.3"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
django-pandas = "^0.6.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
